refactor: extract form that Creates card #103
