+++ src/pages/Deck/subcomponents/CreateCard/index.ts
@@ -1,0 +1,1 @@
+export * from './CreateCard';
