+++ src/pages/Deck/subcomponents/index.ts
@@ -1,2 +1,3 @@
 export * from './DeckList';
 export * from './SectionTitle';
+export * from './CreateCard';
